Use imitation round robin to reduce LOC
